requirements.txt
numpy
nibabel
tensorflow
scikit-learn
scipy
matplotlib
scikit-image
pillow
tk